Don't split lines on \ for directives
There are some consequences to this. Firstly, this works:

Before this change, it would split off the .dw xxxx and produce an
unknown symbol error (unless xxxx was defined).

The other consequence is that this will not work as expected:

.db 10 \ .db 20

This will be evaluated as a single db directive that is being passed in
the arguments ["10", "\", ".db", "20"].

I think this is how it should work but I'm open to more feedback.